+++ python/aistore/sdk/client.py
@@ -36,7 +36,7 @@ class Client:
     Args:
         endpoint (str): AIStore endpoint.
         skip_verify (bool, optional): If True, skip SSL certificate verification. If False (default),
-            the 'AIS_SKIP_VERIFY' environment variable is also checked.
+            the 'AIS_SKIP_VERIFY_CRT' (or legacy 'AIS_SKIP_VERIFY') environment variable is also checked.
         ca_cert (str, optional): Path to a CA certificate file for SSL verification. If not provided,
             the 'AIS_CLIENT_CA' environment variable will be used. Defaults to None.
         client_cert (Union[str, Tuple[str, str], None], optional): Path to a client certificate PEM file